Replace old tick trade view
