+++ src/_selectors/TickTradeSelector.js
@@ -3,13 +3,16 @@ import { fullTradesSelector } from './FullTradeSelector';
 import { findIfExist } from '../_utils/ObjectUtils';
 
 const tickAssetFilter = (assets, contracts) => {
-    const tickAssets = assets.filter(a => {
-        const assetContract = contracts[a];
-        if (!assetContract) {
-            return true;
-        }
-        return findIfExist(assetContract, o => Array.isArray(o) && o.unit === 't');
-    });
+    const tickAssets = Object.keys(assets).reduce((a, b) => {
+        const temp = assets[b].filter(s => {
+            const assetContract = contracts[s];
+            if (!assetContract) {
+                return true;
+            }
+            return findIfExist(assetContract, o => Array.isArray(o) && o.unit === 't');
+        });
+        return a.concat(temp);
+    }, []);
 
     return tickAssets;
 };
@@ -18,8 +21,8 @@ export const tickTradesSelector = createSelector(
     fullTradesSelector,
     trade => {
         const tickAssets = tickAssetFilter(trade.assets, trade.contracts);
-        const cloneTrade = trade;
-        cloneTrade.assets = tickAssets;
-        return cloneTrade;
+        const refTrade = trade;             // not entirely sure if this will create trouble
+        refTrade.assets = tickAssets;
+        return refTrade;
     }
 );
