+++ src/tick-trade/TickTradeMobile.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { MobilePage } from '../_common';
-import TickTradeContainer from './TickTradeContainer2';
+import TickTradeContainer from './TickTradeContainer';
 
 export default (props) => (
 	<MobilePage>
