+++ src/tick-trade/TickTradeParameters.js
@@ -7,83 +7,73 @@ import { M, NumberPlain } from '../_common';
 
 export default class TickTradeParameters extends React.Component {
 
-	shouldComponentUpdate = shouldPureComponentUpdate;
+    shouldComponentUpdate = shouldPureComponentUpdate;
 
-	static propTypes = {
-		assetName: PropTypes.string.isRequired,
-		actions: PropTypes.object.isRequired,
-		assets: PropTypes.object.isRequired,
-		tickTrade: PropTypes.object.isRequired,
-	};
+    static propTypes = {
+        assets: PropTypes.array.isRequired,
+        currency: PropTypes.string.isRequired,
+        durationChange: PropTypes.func,
+        id: PropTypes.string.isRequired,
+        trade: PropTypes.object.isRequired,
+    };
 
-	getPrice() {
-		const { actions, tickTrade } = this.props;
-		actions.subscribeToPriceProposal(tickTrade);
-	}
+    render() {
+        const { currency, durationChange, id, trade } = this.props;
+        const tradeTypeText = tradeToFriendlyType(trade.type, trade.barrier);
 
-	componentDidMount() {
-		this.getPrice();
-	}
-
-	render() {
-		const { actions, assetName, tickTrade } = this.props;
-		const tradeTypeText = tradeToFriendlyType(tickTrade.get('tradeType'), tickTrade.get('barrier'));
-
-		return (
-			<div className="trade-parameters">
-				<div className="row">
-					<fieldset>
-						<label>
-							<M m="Asset" />
-						</label>
-						<br />
-						<Link to={'/asset-picker?goback&tick'} className="btn-secondary">
-							{assetName}
-						</Link>
-					</fieldset>
-					<fieldset>
-						<label>
-							<M m="Type" />
-						</label>
-						<br />
-						<Link
-							className="btn-secondary"
-							to="/trade-type-picker"
-						>
-							<M m={tradeTypeText} />
-						</Link>
-					</fieldset>
-				</div>
-				<div className="row" style={{ marginTop: '1rem' }}>
-					<fieldset>
-						<label>
-							<M m="Ticks" />
-						</label>
-						<RangeGroup
-							min={5} max={10}
-							items={['5', '6', '7', '8', '9', '10']}
-							value={+tickTrade.get('duration')}
-							onChange={e => {
-								actions.updateTickTradeParameters({ duration: e.target.value });
-								this.getPrice();
-							}}
-						/>
-					</fieldset>
-					<fieldset>
-						<label>
-							<M m="Amount" />
-						</label>
-						<br />
-						<Link
-							className="btn-secondary"
-							to="/payout-picker"
-						>
-							<M m={tickTrade.get('basis')} /><span>: </span>
-							<NumberPlain currency={tickTrade.get('currency')} value={tickTrade.get('amount')} />
-						</Link>
-					</fieldset>
-				</div>
-			</div>
-		);
-	}
+        return (
+            <div className="trade-parameters">
+                <div className="row">
+                    <fieldset>
+                        <label>
+                            <M m="Asset" />
+                        </label>
+                        <br />
+                        <Link to={`/asset-picker/${id}`} query={{ type: 'tick' }} className="btn-secondary">
+                            {trade.symbol}
+                        </Link>
+                    </fieldset>
+                    <fieldset>
+                        <label>
+                            <M m="Type" />
+                        </label>
+                        <br />
+                        <Link
+                            className="btn-secondary"
+                            to={`/trade-type-picker/${id}`}
+                        >
+                            <M m={tradeTypeText} />
+                        </Link>
+                    </fieldset>
+                </div>
+                <div className="row" style={{ marginTop: '1rem' }}>
+                    <fieldset>
+                        <label>
+                            <M m="Ticks" />
+                        </label>
+                        <RangeGroup
+                            min={5} max={10}
+                            items={['5', '6', '7', '8', '9', '10']}
+                            value={+trade.duration}
+                            onChange={durationChange}
+                        />
+                    </fieldset>
+                    <fieldset>
+                        <label>
+                            <M m="Amount" />
+                        </label>
+                        <br />
+                        <Link
+                            className="btn-secondary"
+                            to={`/payout-picker/${id}`}
+                            query={{ currency }}
+                        >
+                            <M m={trade.basis} /><span>: </span>
+                            <NumberPlain currency={currency} value={trade.amount} />
+                        </Link>
+                    </fieldset>
+                </div>
+            </div>
+        );
+    }
 }
